requirements.txt
Flask
Flask-smorest
Flask-SQLAlchemy
Flask-JWT-Extended
python-dotenv
uwsgi
psycopg2